Add VR-Tek driver
This is the first cut at a driver for the VR-Tek WVR3. The distortion
coefficients and aberration coefficients still need work, but
rotational tracking is working and the opengl and simple examples can
be used. It also works with SteamVR via SteamVR-OpenHMD, albeit with
correction right now.

There is some early work for the VR-Tek WVR2 also. Rotational tracking
works, thanks to it being the same as for the WVR3, but the display is
rotated 90 degrees.

VR-Tek is using the USB ID of the Oculus DK1, so we need to ensure that
the rift driver doesn't load mistakenly.

The VR-Tek headsets are from Direkt-tek, sold through Walmart in the
US:
http://direkt-tek.com/#vr

Emdoor VR is the OEM for the headsets, with:
* WVR1 -> SpearI FHD
* WVR2 -> SpearI 2K
* WVR3 -> SpearII
http://vr.emdoor.com/en/products/PC_VR/